Authentication JWTStrategy

+++ src/events/events.controller.ts
@@ -1,25 +1,27 @@
+import { ForbiddenException, UnauthorizedException } from '@nestjs/common';
 import { ListEvents } from './input/list.events';
 import { UpdateEventDto } from './input/update-event.dto';
 import { CreateEventDto } from './input/create-event.dto'; 
 import { Event } from './event.entity'; 
-import { Body, Controller, Delete, Get, HttpCode, Param, Put, Patch, Post, ParseIntPipe, ValidationPipe, Logger, NotFoundException, Query, UsePipes } from '@nestjs/common';
+import { Body, Controller, Delete, Get, HttpCode, Param, Put, Patch, Post, ParseIntPipe, ValidationPipe, Logger, NotFoundException, Query, UsePipes, UseGuards } from '@nestjs/common';
 import { Like, MoreThan, Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { throws } from 'assert';
 import { Attendee } from './attendee.entity';
 import { EventsService } from './events.service';
+import { User } from 'src/auth/user.entity';
+import { CurrentUser } from 'src/auth/current-user.decorator';
+import { AuthGuardJwt } from 'src/auth/auth-guard.jwt';
+import console from 'console';
 
 @Controller('events')
 export class EventController {
 
     private readonly logger = new Logger(EventController.name);
 
-    constructor( 
-        @InjectRepository(Event)
-        private readonly eventRepo: Repository<Event>,
-
-        @InjectRepository(Attendee)
-        private readonly attendeeRepo: Repository<Attendee>,
+    constructor(
+        /*@InjectRepository(Attendee)
+        private readonly attendeeRepo: Repository<Attendee>,*/
 
         private readonly eventsService: EventsService,
 
@@ -32,7 +34,6 @@ export class EventController {
     @UsePipes(new ValidationPipe({ transform: true }))
     async findAll(@Query() filter: ListEvents) {
 
-        //console.log("filter: ", filter);
         //this.logger.log(`Hit the findAll route`);
         
         const events = await this.eventsService.getEventsWithAttendeeCountFilteredPaginated(
@@ -47,6 +48,7 @@ export class EventController {
         return events;
     }
 
+/*
     @Get('/practice')
     async practice() {
         return await this.eventRepo.find({
@@ -129,6 +131,7 @@ export class EventController {
 
         return event;
     }
+*/
 
 
     @Get(':id')
@@ -141,45 +144,55 @@ export class EventController {
     }
 
     @Post()
-    async create(@Body() input: CreateEventDto) {
-        const event: any = {
-            ...input,
-            when: new Date(input.when),
-        };
-        const createdEvent =  this.eventRepo.create(event) ;
-        console.log("createdEvent: ", createdEvent);
-
-        const savedEvent = await this.eventRepo.save(createdEvent) ;
-        console.log("savedEvent: ", savedEvent);
-
-        return savedEvent ;
+    @UseGuards(AuthGuardJwt)
+    async create(
+        @Body() input: CreateEventDto,
+        @CurrentUser() user: User | null, 
+        ) 
+    {
+        return await this.eventsService.createEvent(input, user);
     }
 
     @Patch(':id')
-    async update(@Param('id') id: number, @Body() input: UpdateEventDto) {
+    @UseGuards(AuthGuardJwt)
+    async update(
+        @Param('id') id: number, 
+        @Body() input: UpdateEventDto,
+        @CurrentUser() user: User) {
 
-        let actualEvent = await this.eventRepo.findOneBy({ id: id }); 
+        let actualEvent = await this.eventsService.getEvent(id);
 
         if (actualEvent === null) {
                 throw new NotFoundException();
         }
 
-        let newInput:Event = {
-            ...actualEvent,
-            ...input,
-            when: input.when ? new Date(input.when) : actualEvent.when
-        } ;
+        if (actualEvent.organizerId !==  user.id)  {
+            throw new ForbiddenException(
+                null, `You are not authorized to change this event!`);
+        }
 
-        //let updatedEvent = await this.eventRepo.update({ id }, newInput); 
-        let updatedEvent = await this.eventRepo.save(newInput); 
+        let updatedEvent = await this.eventsService.updateEvent(actualEvent, input);
 
         console.log("updatedEvent: ", updatedEvent)
         return updatedEvent ;
     }
 
     @Delete(':id')
-    @HttpCode(204)
-    async delete(@Param('id') id) {
+    @UseGuards(AuthGuardJwt)
+    //@HttpCode(204)
+    async delete(
+        @Param('id') id,
+        @CurrentUser() user: User) {
+
+        let event = await this.eventsService.getEvent(id) ;
+
+        if (!event) {
+            throw new NotFoundException();
+        }
+
+        if (event.organizerId !==  user.id) {
+            throw new ForbiddenException(null, `You are not authorized to delete this event!`);
+        }
 
         const result =  await this.eventsService.deleteEvent(id);
 
@@ -188,7 +201,5 @@ export class EventController {
         } else {
            return "Deleted with success!"            
         }
-
-        
     }
 }
